pyproject.toml
[project]
name = "goofy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.3",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
